Revert GCC change about turning -Wimplicit-function-declaration into errors
Unfortunately, the new behavior breaks the compilation of some GCC dependencies and even GCC itself when building for very old targets.

This is due to these platforms lacking support for some required standard library functions and APIs: Until GCC 13, the compiler always fell back to a built-in alternative in case one of those functions was missing (see https://gcc.gnu.org/onlinedocs/gcc/Other-Builtins.html). The compiler, however, used to just issue an "implicit function declaration" warning for the usage of those builtins. Since this warning is now treated as an error, this simply made the compilation fail.