Disable emitting .gnu.hash for MIPS targets
The .gnu.hash and the MIPS ABI require .dynsym to be sorted in different ways. .gnu.hash needs symbols to be grouped by hash code, whereas the MIPS ABI requires a mapping between the GOT and the symbol table.

This change was not required until now because we were using the BFD linker by default, and the BFD linker silently ignored the .gnu.hash when targeting MIPS ABIs. Since we are now using the gold linker, we encounter the following errors while linking MIPS code:

$ mips64el-unknown-linux-gnuabi642.24-g++ main.c
ld.gold: error: .gnu.hash is incompatible with the MIPS ABI
collect2: error: ld returned 1 exit status