Replace -f{no-}fat-lto-objects with -flto-partition={none,balanced} in the Clang wrapper
I was wrong in assuming what the -flto={full,thin} flag in Clang was about. The value passed to that argument does not specify whether Clang should generate fat or non-fat LTO objects; instead, it controls whether the compiler should split the LTO objects into multiple partitions (thin) or use a single partition (full).

With this in mind, the equivalent of Clang's -flto=full in GCC should be -flto-partition=none, and -flto=thin should correspond to -flto-partition=balanced.