Support cross-compiling with Clang
This commit adds the necessary logic to allow Clang to be used instead of GCC for cross-compiling. This is done by invoking the system clang/clang++ compiler with the appropriate --sysroot option.

It should be noted that, despite using Clang, any project built with it will be linked against GCC's runtime libraries rather than Clang's. This is because, while Clang has native support for cross-compilation, it does not build the runtime libraries for all of its supported targets by default.