Ignore header files under (prefix)/(system-root)/include-missing
These header files are not meant to be included. They exist to prevent projects from trying to include headers from the host C library when using OBGGCC_SYSTEM_LIBRARIES and the requested header does not exist within the cross-compiler’s system root. This is done so projects do not accidentally use features that are not available in the specific glibc version they are being cross-compiled against.