Enable automatic linking of librt
On glibc 2.16 and earlier, the clock-related routines from the C library are implemented in librt rather than libc. Because of this, you must explicitly link against -lrt in order to use them.

While this is usually not a problem for people building their own projects — where they have direct control over the build system and can easily adjust the CFLAGS/LDFLAGS — many third-party projects do not include the necessary logic to link against this library on older glibc versions. Having to patch each of these projects individually to add that support is annoying.